Move the old UI's AppBar into its own component
